+++ src/main/java/com/recurly/v3/requests/SubscriptionChangeShippingCreate.java
@@ -9,6 +9,7 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.recurly.v3.Request;
 import com.recurly.v3.resources.*;
+import java.math.BigDecimal;
 
 public class SubscriptionChangeShippingCreate extends Request {
 
@@ -18,7 +19,7 @@ public class SubscriptionChangeShippingCreate extends Request {
    */
   @SerializedName("amount")
   @Expose
-  private Float amount;
+  private BigDecimal amount;
 
   /**
    * The code of the shipping method used to deliver the subscription. To remove shipping set this
@@ -42,7 +43,7 @@ public class SubscriptionChangeShippingCreate extends Request {
    * Assigns the subscription's shipping cost. If this is greater than zero then a `method_id` or
    * `method_code` is required.
    */
-  public Float getAmount() {
+  public BigDecimal getAmount() {
     return this.amount;
   }
 
@@ -50,7 +51,7 @@ public class SubscriptionChangeShippingCreate extends Request {
    * @param amount Assigns the subscription's shipping cost. If this is greater than zero then a
    *     `method_id` or `method_code` is required.
    */
-  public void setAmount(final Float amount) {
+  public void setAmount(final BigDecimal amount) {
     this.amount = amount;
   }
