+++ src/main/java/com/recurly/v3/resources/AccountAcquisitionCost.java
@@ -8,13 +8,14 @@ package com.recurly.v3.resources;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.recurly.v3.Resource;
+import java.math.BigDecimal;
 
 public class AccountAcquisitionCost extends Resource {
 
   /** The amount of the corresponding currency used to acquire the account. */
   @SerializedName("amount")
   @Expose
-  private Float amount;
+  private BigDecimal amount;
 
   /** 3-letter ISO 4217 currency code. */
   @SerializedName("currency")
@@ -22,12 +23,12 @@ public class AccountAcquisitionCost extends Resource {
   private String currency;
 
   /** The amount of the corresponding currency used to acquire the account. */
-  public Float getAmount() {
+  public BigDecimal getAmount() {
     return this.amount;
   }
 
   /** @param amount The amount of the corresponding currency used to acquire the account. */
-  public void setAmount(final Float amount) {
+  public void setAmount(final BigDecimal amount) {
     this.amount = amount;
   }
