+++ src/main/java/com/recurly/v3/resources/PlanPricing.java
@@ -8,6 +8,7 @@ package com.recurly.v3.resources;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.recurly.v3.Resource;
+import java.math.BigDecimal;
 
 public class PlanPricing extends Resource {
 
@@ -23,12 +24,12 @@ public class PlanPricing extends Resource {
    */
   @SerializedName("setup_fee")
   @Expose
-  private Float setupFee;
+  private BigDecimal setupFee;
 
   /** Unit price */
   @SerializedName("unit_amount")
   @Expose
-  private Float unitAmount;
+  private BigDecimal unitAmount;
 
   /** 3-letter ISO 4217 currency code. */
   public String getCurrency() {
@@ -45,7 +46,7 @@ public class PlanPricing extends Resource {
    * cycle. For subscription plans with a trial, the setup fee will be charged at the time of
    * signup. Setup fees do not increase with the quantity of a subscription plan.
    */
-  public Float getSetupFee() {
+  public BigDecimal getSetupFee() {
     return this.setupFee;
   }
 
@@ -55,17 +56,17 @@ public class PlanPricing extends Resource {
    *     charged at the time of signup. Setup fees do not increase with the quantity of a
    *     subscription plan.
    */
-  public void setSetupFee(final Float setupFee) {
+  public void setSetupFee(final BigDecimal setupFee) {
     this.setupFee = setupFee;
   }
 
   /** Unit price */
-  public Float getUnitAmount() {
+  public BigDecimal getUnitAmount() {
     return this.unitAmount;
   }
 
   /** @param unitAmount Unit price */
-  public void setUnitAmount(final Float unitAmount) {
+  public void setUnitAmount(final BigDecimal unitAmount) {
     this.unitAmount = unitAmount;
   }
 }
